requirements.txt
# Core
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1

# Database
sqlalchemy==2.0.36

# Email / Gmail API
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.125.0

# Scheduling
apscheduler==3.10.4

# Web Dashboard
flask==3.0.3
flask-wtf==1.2.1
werkzeug==3.0.2

# Templating / utilities
jinja2==3.1.3
python-dotenv==1.0.1
pyyaml==6.0.1
click==8.1.7
rich==13.7.1

# AI Personalization
google-generativeai>=0.8.0
